Correctly normalize parameters for oAuth authentication
Currently oAuth authenfication for nested arrays is broken as those
information is not encoded like the server decodes the information.
This patch simplifies the normalization handling and processes
nested arrays in a proper way. Unfortunately oAuth documentation
fails to give some expectations on how to normalize nested array
information correctly.